Skip function entry points
The perfparser gives us frames for function entry points, i.e. the
list of inlined frames usually has for code such as this:

void foo() // one frame here
{
  ...
  actualCostlyStuff; // one frame here
  ...
}

For our purposes, the inner frame is the interesting one and the
outer one can be ignored. This patch does this now.