Heavily refactor the models and associated data structures
This is a big change that lays the foundation for some important
features, which will improve the usability of hotspot:

- always aggregate by [symbol, binary] pair
- add lists of callers and callees to the caller/callee map
- an "annotate source" feature

The central ideas are the following:

- Data::Symbol is the unique pair of symbol name and binary name
- Data::Location is the unique pair of an address and resolved
  file/line location
- the caller/callee map aggregates all the information for a given
  Data::Symbol. This includes:
  - total inclusive cost
  - total self cost
  - all locations within this symbol with separate inclusive/self cost
  - all callers, i.e. other Data::Symbol and Data::Locations that
    called this Data::Symbol
  - all callees, i.e. other Data::Symbols that are being called from
    a given Data::Location within this Data::Symbol

In order for this to be implemented properly, a unit test framework
is added that takes string inputs and builds trees out of it, which
can then again be converted to strings for comparisons against the
expectations.

This also uncovered some bugs in the algorithms to build the top-down
and caller/callee data sets from the bottom-up data. This will be
fixed in follow-up commits.