Fix building the top-down data tree
Our aggregated bottom-up tree can contain "partial leafs". That is,
nodes for which the sum of the children's cost is less than the nodes
cost itself. This means, that this node has some direct self cost.
Such cases must also be handled like "real leafs" who don't have any
children.

Additionally, we have to always use the aggregated cost when
incrementing the self-cost. Otherwise we get wrong numbers when we
encounter an aggregated stack in the bottom-up tree.