Forward sysroot, kallsyms, extra libs and debug paths to perfparser
With these new CLI args, hotspot can be used for cross-machine
reporting of perf.data files. E.g.:

On an armv7 machine, record some data:

    perf record --call-graph dwarf ...

Once finished, copy over the perf.data file and /proc/kallsyms.
Then run hotspot and pass the path to the sysroot, which hopefully
contains the debug information necessary for unwinding and symbol
resolution:

    hotspot --sysroot ... --kallsyms kallsyms perf.data

In the future, we will need to let people specify the values of
these settings from the GUI.