Report per-attribute sample counts in summary
This is a first step towards task #14, i.e. supporting perf.data files
with multiple event sources. But this also depends on a not-yet merged
fix to upstream perfparser, namely:

https://codereview.qt-project.org/#/c/189579/

With that one applied, we will now see the following in the summary
for a perf.data file recorded with `-e cycles,instr,cache-misses`:

Total Samples: 3521
    cycles:u: 1743
    cache-misses:u: 37
    instructions:u: 1741