Use github.com/KDAB/perfparser as source for perfparser submodule
This allows us to integrate some not-yet upstreamed features into
hotspot without having to wait for these changes to have gone
through code review upstream.

This is of course only a temporary measure until all the required
features are upstreamed. Until then, this already allows us to
parse the sample weights and periods, which we can make use of
in follow-up commits.

Additionally, we already get access to the much improved performance
of perfparser that I've been working on recently.