Use QMenuBar::setCornerWidget instead of setStyleSheet for KDAB icon
Using QSS on the mainwindow messes up the breeze style - it won't
paint the busy indicator of a progress bar anymore e.g.

Using the corner widget API is also cleaner and allows us to
add a click handler to open the KDAB website. We can now also add
a tooltip here.