Optimize HashModel
Do not iterate linearly over the underlying QHash as that is
very slow. Especially when we profile a large application with
thousands of different symbols, the calls to std::distance became
a big hotspot when the caller/callee model was sorted.

Now, we drop the hash and store the data instead in two vectors for
the keys and values. This dramatically improves the runtime
performance of the HashModel for large hash sets, at the cost of
more memory consumption as we duplicate the data.