Decouple caller or callee view navigation from callerCalleeView
The latter view may be filtered. In such cases, a caller or callee
may not be visible, and thus have no valid proxy index. In such cases,
the navigation would break.

Now we have decoupled this feature from the callerCalleeView. It
can be filtered, but navigation in the caller/callee views will still
work as expected.