Visualize data from perf record runs with multiple events
The parser and models now expect a dynamic number of cost columns,
defined at runtime based on the attributes found in the perf data
file. This allows us to visualize data for multiple metrics in one
central view, which is much more comfortable than having to switch
between different report "tabs" as is done in `perf report`.

The code does become more complicated, as I tried to minimize the
overhead. I.e. instead of having a dynamically sized array in every
model entry, we use a shared Costs data structure, which contains
the cost values for all metrics.

Fixes: https://github.com/KDAB/hotspot/issues/14