Display relative costs in tables
This resembles perf report's overhead column better. Showing the raw
aggregated costs in the tables is imo misleading, as we do sampling
and only want to show statistical distributions here. The raw numbers
of e.g. cycles make it sound as if a function really consumed X cycles,
measured with high accuracy (which is not the case).