Include the dependencies of hotspot-perfparser in the AppImage
We also need to extract the library dependency for hotspot-perfparser.
Just calling linuxdeployqt again seems to be enough. It will copy
the dependencies of the perfparser into its own lib folder and setup
RPATH correctly, which is fine.

In the future we could try to share common libs, by manually moving
the libraries and symlinking one lib folder to the other.

Fixes: https://github.com/KDAB/hotspot/issues/87