Move ebl libs into the same folder where libdw.so.1 resides
This ensures that these libs are properly found from the dlopen
call inside libdw. The previous location inside elfutils/ would
not work properly on systems such as OpenSuse or Ubuntu, where
libs are only found in special folders. I.e. on OpenSuse it would
look for lib64/elfutils, and on Ubuntu for lib/x86_64-linux-gnu.
Only on Arch did the previous solution work properly for me.
By putting the libs directly into the folder of libdw, it seems to
work on all of these platforms.

Fixes: https://github.com/KDAB/hotspot/issues/88