Share libraries between hotspot and hotspot-perfparser in AppImage
I hoped that this would reduce the size of the AppImage a bit, but
it doesn't - probably due to the smartness of squashfs. That said,
it's still the sensible thing to do imo. It could also hopefully
lead to a (small) memory consumption reduction at runtime, as the
same share libraries can be reused for the perfparser executable
that are also used by hotspot itself.