Refactor zooming code to operate on time ranges directly
This cleans up the code in a few places. But, more importantly,
it also fixes the strange behavior when resizing the timeline
column: Previously, this would change the zoomed range, as the size
of the column would influence the selected times. By mapping the
point to time directly, we can resize the column however we want
and it still does the correct thing (i.e. it will always show
the same time slice).