Use formlayout and stacked widget for the record page
The form layout is prettier for forms. But empty form layout cells
are not completely collapsed, something that needs to be fixed
somewhere upstream. So instead, use a QStackedWidget for the launch/
attach form widgets. This also simplifies the code a bit, but
requires us to use a KColumnResizer to keep the form column widths
in sync. Additionally, it requires us to make the stack widget use
only the needed size, which adds code complexity again...

Overall, somewhat more complicated code, but the result is prettier.