Document failing test that uncovers unwinding issues for cpp-recursion
Unwinding from fibonacci often fails, but this is also happening
with perf script using both, libunwind or libdw for unwinding.
Increasing the stack size doesn't help either. It's not clear to
me what the issue is. I'll have to report it upstream in the hope
to get someone else to fix this.