Rework the start page layout
Place the open file + settings buttons in one row, and the record
button in a row below. The settings only apply to opening a file
after all. When we launch a new record, the defaults should be
fine and no settings need to be set.

Also ensure that the settings button and other buttons have the
same height. The record button on the other hand has the same width
as the file open button and settings paired together.