Update perfparser submodule
This includes an improvement to the error reporting for missing
elf files. Most notably, we won't get visible errors reported
for non-elf files that got removed in the meantime. This probably
also helps with issue #44, as the deleted mappings should never
contain any executable code to begin with, or so I hope.