Don't report error about missing debug symbols in unknown modules
When we fail to map an IP to an ELF, we show an error for that
already. That of course means we cannot resolve its symbol, and then
later we would report a second error about missing debug symbols
in the module "" (empty string).

This patch removes this second error.