Support off-CPU profiling
When we encounter a perf data file that was recorded with

    perf record --switch-events -e sched:sched_switch --call-graph dwarf

then we now take the off-CPU time from the switch events and
the call stack from the sched_switch events to support off-CPU
profiling. I.e. we will show both the number of scheduler switches
in the sched:sched_switch column, as well as the off-CPU time in
a new column.

In the timeline, the off-CPU time is always highlighted in red and
can be hovered to inspect the sched switch delay.

Fixes #8