Try to find source file relative to module path
Some binaries use relative paths in the debug information. Most
notably, this happens when you build with `qmake ../...` which
I do quite frequently. In these situations, we often failed to
find the source file, despite it being available.

Now also look for the file relative to the module path, which
fixes the above issue.