Optionally enable rustc_demangle when user sets include+lib paths
To enable, the user has to set RUSTC_DEMANGLE_INCLUDE_DIR and
RUSTC_DEMANGLE_LIBRARY when compiling hotspot. You can use ccmake
or a cmake-gui too to set the cache variables. The include dir
needs to point to the path that contains rustc_demangle.h and
the library needs to point at the librustc_demangle.so. Both
can be obtained from https://github.com/alexcrichton/rustc-demangle.

Fixes: https://github.com/KDAB/hotspot/issues/237