Report number of missing/total symbols per module
This change extends the information contained in the error message
displayed to the user when a module is missing debug symbols.

Suppose that, in the recorded samples, there are 40 locations in module
`ld-2.30.so`, two of which could not be resolved to a symbol. Then a
message will be displayed in under _Summary/Parser Errors_.

Message before this change:

    Module "ld-2.30.so" is missing (some) debug symbols.

Message after this change:

    Module "ld-2.30.so" is missing 2 of 40 debug symbols.

This additional information should be sufficient to help the user
determine whether a relevant amount of debug symbols are missing or not.
So, IMHO this resolves #54