Added option --branch-traverse to traverse by option only branchStack in resolveCallchain for lbr.
For the moment there are only branchStack items are traversed for LBR.
It is not enough to compute event costs by instructions for Disassembly.
Added more frames to traverse. But it can have a certain impact to performance for LBR.
Entered option 'branch-traverse' to turn off traverse by additional frames for LBR.
When it is passed the only branchStack is traversed and events costs are not showed.