Implement exporting of raw perfparser data
This is the first step in allowing users to convert perf.data into
portable perfparser files that we can then open on other machines
for analysis, without requiring us to redo the full unwinding nor
access to debug symbols and binaries.

The format of the exported data is not stable, and as such this
is not (yet) meant backup purposes.