Apply hover effect to stacks from event timeline to flame graph
This makes the two views behave more in sync. I.e. we can now
hover a frame in the flame graph and see all events that have that
symbol highlighted in the timeline - that existed for some time now.

This patch adds the other way around: highlighting the stack in the
flame graph for the event we hover in the timeline delegate.