Rework CI and docker images to allow better caching
Split the docker files apart:

- a base image which is distro specific
- the BuildDependencies one which, for now, installs KDDW
- the BuildHotspot one which either builds using cmake presets or
  just normally with plain CMake

Then allow us to publish the base image + dependencies to ghcr.io
and pull those in the github workflow then. This greatly improves
the build times for the CI, as we don't need to wait for the docker
images to be build anymore, and instead can concentrate on hotspot.

While at it, also run the compile-and-test workflow for pull requests
and allow it to be manually dispatched too.