Keep custom cost aggregation groups on top
When we build the top-down graph from the bottom-up graph that
uses a custom cost aggregation, we used to put the group (e.g.
the thread name etc.) at the bottom. This isn't desired, it's much
more useful to also split the top-down data by that group on the
top, which this patch here implements.

In order for this to work reliably, we also have to take these
new top-groups into account when we want to highlight stacks.