Add scripts to generate docker baseimage for appimage on github action
These files here build the foundation for creating an appimage on
a github action. With them, we can create a docker image manually
with a centos 7 setup and a custom Qt 5.15.7 and KF5 build from
sources. Then we can upload it to ghcr.io and leverage it by the
github actions. These will then "just" download the image, which
is relatively fast compared to compiling everything from scratch.

Note that we do not yet build konsolepart, as it requires a lot
of additional dependencies. Maybe we revise that decision in the
future.

Additionally, kgraphviewer crashes when we include it in the
appimage, so we don't do that either. Something else to look into
in the future.

The appimage created this way includes the other useful optional
dependencies though - QCustomPlot and rust + d demangling.