Introduce Data::FileLine
Instead of allocating QStrings for every file+line combination,
keep the data separate to better exploit COW semantics of QString.
This reduces the number of memory allocations during parsing
drastically.

Furhtermore, we will be able to use this data now more easily for
the source code model in the disassembly view, which will be done
in a follow-up patch.