Leverage sourceMap instead of offsetMap for SourceCodeModel
This way we get a proper attribution of all costs at a given
line of code, esp. for inclusive costs. The offset map was not
always hitting all code paths and did not yield correct costs that
matched what we showed in the source map view of the caller/callee
model.

Fixes: https://github.com/KDAB/hotspot/issues/427