Use system monospace font disassembly and source code views
Instead of setting "monospace" in the .ui file on the view, which
would then be picked up by the header view too, leverage the model
to decide what font to use based on the FontRole. This allows us
to only use the fixed width font for the source code and disassemmbly
columns but not the line number of cost columns.

This also uncovered that we didn't actually set the fixed width
font on the highlighter, which might potentially have lead to
bad performance when layouting a lot of content.

Relates-To: https://github.com/KDAB/hotspot/issues/428