Don't mutate highlighted document in source code model
This sometimes breaks highlighting further down in the source code
as shown in the dummy.c example in issue #428. Instead, we now just
pass the pretty symbol along for the first row in the model, bypassing
highlighting for it too which is imo fine as this line is "special"
and more related to the function indicators we have for inline code
in the dissassembly model.