Opt-in via CMake to allow privilege escalation
Matthias Gerstner from Suse found a TOCTOU security vulnerability
in our privilege escalation code. If abused, the temporary escalation
would remain permanent and opens the door for attackers to wreak
havoc. As this is an optional feature, let's disable it by default
until I find the time to resolve this differently.

See CVE-2023-28144 for more information, the gist is, to quote
Matthias:

> The script contains the following logic during early startup:
>
>     if [ ! -z "$1" ]; then
>         olduser=$(stat -c '%u' "$1")
>         chown "$(whoami)" "$1"
>         echo "rewriting to $1"
>         # redirect output to file, to enable parsing of output even when
>         # the graphical sudo helper like kdesudo isn't forwarding the text properly
>         $0 2>&1 | tee -a "$1"
>         chown "$olduser" "$1"
>         exit
>     fi
>
> The two `chown` invocations on the temporary file result in a temporary change
> of the ownership of the temporary file to root, which is originally owned by
> the unprivileged user. So it changes ownership of the provided path first to
> `root`, then reexecutes itself, then changes ownership back to the original
> user.
>
> This offers the following attack vectors:
>
> - giving ownership of an arbitrary file to root
> - giving ownership of an arbitrary file to the unprivileged user

Thanks a lot for Matthias and Suse for detecting this issue and
bringing it to my attention.