Hide empty tracepoint cost columns in the dissassembly dock
For these costs, we would never have self costs on an instruction
level they only make sense in an "inclusive" way, which we don't
show in the instruction level assembly view.

Note that the `hideEmptyColumns` isn't enough as that works on
the total costs level. We would want to filter on the costs that
are shown in the respective model instead, but that's not what these
function calls are doing.