Implement new perf elevated privilege system
Instead of using a script that changes system configuration
temporarily, just run perf as root.

To properly synchronize with a launched app (that itself should not
run as root), launch the app in a separate, initially stopped process,
and use the control fifo feature of perf to properly synchronize with
it.

The control fifos are also needed to be able to stop sudo-perf, as one
does not have permission to SIGINT it anymore.