Lay the foundations to compile against Qt6
When we pass `cmake -DQT6_BUILD=ON ...` then we will try to find
Qt6 and KF6 instead of Qt5 and KF5 and build against that. This patch
also removes the version 5 from some config flags to make the code
more portable.

Note that the Qt6 config won't compile yet, this will be fixed in
follow-up patches.