Improve debuginfod download progress reporting
Show the module name for which the debug information is being
downloaded and only show the URL in the tooltip - it's usually
uninteresting for users. But do show the size of the file that is
being downloaded and how much data was downloaded, instead of only
showing the percentage value next to the progress bar.