shove flamegraph controls into QToolBar and reorder them slightly
This change removes the minimum width requirements of about
1100pixels for the hotspot mainwindow. The remaining window size
constraints are defined by the tabs. When the window is small,
the QToolBar will hide the actions and show a button to list them
in a drop down menu. We also reorder the actions slightly and
give the search box a minimum width, to make it more useable.

fixes: #501