Declare `asmProcess` after `output` to ensure it's always valid
When `readyRead` is emitted on destruction of the `QProcess`, we need
to be able to access a fully constructed `output` still. Previously
this could trigger crashes apperently, which should now be resolved.

I do have to admit though, that I don't quiet get when this actually
can happen - `waitForFinished` should, or so I thought, drain the
process buffer already...

Fixes: https://github.com/KDAB/hotspot/issues/542