Cleanup perf record and control fifo after finish
When we elevate privileges and control perf through the FIFO we
should cleanup the processes once the recording is finished.
Otherwise, the next time a record is started we would request
stop on the FIFO which would fail at that stage then, but recording
actually worked just fine - only because we forgot to cleanup the
processes already earlier.