Add error message when exporting a flamegraph that hasn't been opened
In order to save a flamegraph to an image or SVG, it must have been
opened at least once so the graphics item is loaded. If you try to
export it before doing so, you'll still get prompted to save the file,
but it won't be (nor can be) written.

Now there's an explicit check with it's own error message in that case.