build appimage with static runtime
This allows the user to run the appimage on non fuse2 systems (all
moderns systems use fuse3). This is achieved by using the static
appimage runtime. Since linuxdeploy doesn't support changing the runtime
I only use it to create the appdir directory. Then I use the original
appimagetool to package the appimage using the static runtime.

This should also make it possible to run the appimage on a non glibc
system but I don't have access to one so I can't test it.

fixes: #485