Fix appimage docker file generation, esp. make kdesrc-build run
We need to skip gpgme and build a kgraphviewer branch that's
compatible with KF5. Then we also modernize the kdesrc-build
usage in general, allowing us to set the tag option on the
frameworks module-set directly. That means we can get rid of
the custom kf5-frameworks-build-include copy and use the upstream
metadata definition directly