fix: crash in disassembler on hover
When hovering the disassembler tries to create an tooltip. If the user
hovers on the disassembly column or an earlier one, the model calculates
a negative cost type which will cause an out of bounds access error in
the cost array, even if the cost is not shown.
This patch changes the order of operations and adds an check to make
sure no out of bound access happens.
fixes: #642