fix: correct cost attribution of inline frames in disassembly view
Previously, we put the cost of an inline frame into a container
indexed by the inline symbol. But during disassembly, we never get
to query that data again, since we cannot disassemble an inline frame.
Instead, we need to be able to query the cost for arbitrary binary
offsets, independent of their originating symbol.

The patch here achieves this by lifting the OffsetLocationCostMap out
of the CallerCalleeEntryMap into CallerCalleeResults, but mapped by
the binary name. This way we can efficiently store and lookup the data
of a given offset within a specific binary during disassembly, which
allows us to show the cost for inlined code in the disassembly view.

Fixes: https://github.com/KDAB/hotspot/issues/671