Introduce a new "By File" cost aggregation tab
Here we look at the debug information to attribute costs to
individual source files. As such, this is similar to the caller/callee
view, but just on a per-file basis.

This is useful for complex projects where you would like to get
an overview on individual subsystems, where we usually have multiple
symbols grouped in one file.

Fixes: https://github.com/KDAB/hotspot/issues/665