fix: don't try to use a cross-architecture objdump on the native architecture
If the host architecture is aarch64, aarch64-linux-gnu-objdump
may not exist, which is the case on at least Fedora and NixOS, so
disassembly doesn't work by default. Similar logic likely also applies
to arm-linux-gnueabi-objdump. Fix the problem by only trying to use the
cross-architecture objdump if it is found in $PATH.