fix: avoid OOB access if source code unreadable
We set m_startLine and m_numLines according to the debug info but
if the source file is unreadable, or if it is readable but does not
have enough lines, this will lead to an OOB access and a crash in
SourceCodeModel::data() -> HighlightedText::lineAt(). Fix it by making
it so that we only set m_startLine and m_highlightedText once we have
successfully read the file and ensure that they are in bounds of the
lines array.